refactor scenarios
