+++ packages/azure-http-specs/specs/client/structure/api-version/query/client.tsp
@@ -1,0 +1,13 @@
+import "@typespec/http";
+import "@typespec/spector";
+import "@azure-tools/typespec-client-generator-core";
+import "./main.tsp";
+
+using TypeSpec.Http;
+using Azure.ClientGenerator.Core;
+using Client.AlternateApiVersion.Service;
+using Spector;
+
+namespace Customizations;
+
+@@apiVersion(QueryApiVersionParam.version);
